Add files via upload
I had trouble getting these to commit, so I had to do it this way. There are two JPEG images needed in the report which I could not get to read and knit in the final report template used. Dr. Li and Dr. Brown did not know the answer to how to get it to work. I spent 10 or more hours trying so you will have to read the images in the word doc I sent to your email.